Incorporate changes from __future__ into 2015 code
Added classes to govern trajectory following based on physical
modelling of drive paths and trajectories. Correspondingly, a
synchronous PID controller was necessary to allow for calculation
of turning PID in real-time instead of asynchronously. Necessary
constants in RobotMap need to be tuned. If requested, a back-end
can be written which allows for DriveStraight continuing a
previous motion instead of having to come to a stop before driving
forwards. See you in the future~~~